requirements.txt
pyqt6
torch
diffusers
accelerate
transformers
pillow
sentencepiece
protobuf